Enhance Simplifier to minimize child PassThroughs
This change adds a check in the simplifier helper logic to see if
a child, which has been simplified, has become a PassThrough. If
the root node does not answer true to isNullChk() then attach the
child of the PassThrough to minimize the number of PassThrough
nodes and maximize opportunities for optimization.

Signed-off-by: Andrew Craik <ajcraik@ca.ibm.com>