Fix z gprClobberEvaluate to ensure new register is flagged as collected reference
The method gprClobberEvaluate may allocate a new register. In this case the new
register may need to be flagged as a collected reference. This commit adds code
code which calls `setContainsCollectedReference` on the destination register
check if the source register contains a collected referece.

Issue: https://github.com/eclipse/openj9/issues/8397

Signed-off-by: Ryan Shukla <ryans@ibm.com>